# server/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Authenticated LXD REST API
host: 192.168.50.4:9000
basePath: /v1
schemes:
  - http
  - https
securityDefinitions:
  token:
    description: The web token required for authorization
    type: apiKey
    name: token
    in: query
consumes:
  - application/json
produces:
  - application/json
paths:
  /authenticate:
    x-swagger-router-controller: auth
    post:
      description: Authenticate a user session and return the auth token
      operationId: signin
      parameters:
        - name: user
          in: body
          description: User
          required: true
          schema:
            $ref: '#/definitions/UserModel'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AuthResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /serverinfo:
    x-swagger-router-controller: serverinfo
    get:
      security:
        - token: []
      description: Returns server information
      operationId: getServerInfo
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ServerInfoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /containers/{name}:
    x-swagger-router-controller: containers
    post:
      security:
        - token: []
      description: Updates a container config
      operationId: updateContainer
      parameters:
        - name: name
          in: path
          description: Container name
          required: true
          type: string
        - name: container
          in: body
          description: Container config data
          required: true
          schema:
            $ref: '#/definitions/ContainerConfigModel'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CreateResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Rename a container
      operationId: renameContainer
      security:
        - token: []
      parameters:
        - name: name
          in: path
          description: Container name
          required: true
          type: string
        - name: container
          in: body
          description: Container data
          required: true
          schema:
            $ref: '#/definitions/ContainerModel'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CreateResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: Returns container information
      operationId: getContainer
      security:
        - token: []
      parameters:
        - name: name
          in: path
          description: Container name
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ContainerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /containers:
    x-swagger-router-controller: containers
    get:
      description: Returns a list of containers
      operationId: getAllContainers
      security:
        - token: []
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ContainersResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Creates a new container
      operationId: createContainer
      security:
        - token: []
      parameters:
        - name: container
          in: body
          description: Container data
          required: true
          schema:
            $ref: '#/definitions/ContainerModel'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CreateResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /containers/detailed:
    x-swagger-router-controller: containers
    get:
      description: Returns a list of containers with extra information
      operationId: getAllContainersWithDetails
      security:
        - token: []
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ContainersResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /profiles:
    x-swagger-router-controller: profiles
    get:
      description: Returns a list of profiles
      operationId: getAllProfiles
      security:
        - token: []
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ProfilesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /profiles/detailed:
    x-swagger-router-controller: profiles
    get:
      description: Returns a list of profiles with extra information.
      operationId: getAllProfilesWithDetails
      security:
        - token: []
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ProfilesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /images:
    x-swagger-router-controller: images
    get:
      description: Returns a list of images.
      operationId: getAllImages
      security:
        - token: []
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ImagesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Create a new image
      operationId: createImage
      security:
        - token: []
      parameters:
        - name: image
          in: body
          description: Image data
          required: true
          schema:
            $ref: '#/definitions/ImageCreateModel'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CreateResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /images/{fingerprint}:
    x-swagger-router-controller: images
    delete:
      description: Deletes an image.
      operationId: deleteImage
      security:
        - token: []
      parameters:
        - name: fingerprint
          in: path
          description: Image fingerprint
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /images/remote/{name}:
    x-swagger-router-controller: images
    get:
      description: Returns a list of remote images for a given server.
      operationId: getRemoteImages
      security:
        - token: []
      parameters:
        - name: name
          in: path
          description: Remote server name
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RemoteImagesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /remote:
    x-swagger-router-controller: servers
    get:
      description: Returns a list of remote image servers.
      operationId: getServers
      security:
        - token: []
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ServersResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /images/detailed:
    x-swagger-router-controller: images
    get:
      description: Returns a list of images with extra information.
      operationId: getAllImagesWithDetails
      security:
        - token: []
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ImagesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /images/aliases:
    x-swagger-router-controller: images
    post:
      description: Create a new alias
      operationId: createImageAlias
      security:
        - token: []
      parameters:
        - name: alias
          in: body
          description: Alias data
          required: true
          schema:
            $ref: '#/definitions/ImageAliasCreateModel'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CreateResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: Returns a list of image aliases.
      operationId: getAllImageAliases
      security:
        - token: []
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ImageAliasesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger.json:
    x-swagger-pipe: swagger_raw
definitions:
  UserModel:
    required:
      - username
      - password
    properties:
      username:
        description: "User username"
        type: string
        minLength: 1
      password:
        description: "User password"
        type: string
        minLength: 1
  ImageAliasCreateModel:
    required:
      - description
      - target
      - name
    properties:
      description:
        type: string
        minLength: 1
      target:
        type: string
        minLength: 1
      name:
        type: string
        minLength: 1
  ContainerModel:
    required:
      - name
      - profiles
    properties:
      name:
        type: string
        minLength: 1
      profiles:
        type: array
        items:
          $ref: "#/definitions/ProfileModel"
        minItems: 1
      image:
        type: object
        $ref: "#/definitions/ImageModel"
  ProfileModel:
    required:
      - name
    properties:
      name:
        type: string
  ImageAliasModel:
    required:
      - alias
    properties:
      alias:
        type: string
  ImageModel:
    required:
      - fingerprint
    properties:
      fingerprint:
        type: string
  RemoteImageModel:
    required:
      - alias
      - fingerprint
      - public
      - description
      - arch
      - size
      - upload_date
    properties:
      alias:
        type: string
      fingerprint:
        type: string
      public:
        type: boolean
      description:
        type: string
      arch:
        type: string
      size:
        type: string
      upload_date:
        type: string
      value:
        type: string
      label:
        type: string
  ServerModel:
    required:
      - name
    properties:
      name:
        type: string
  ImageCreateModel:
    required:
      - public
      - source
    properties:
      public:
        type: boolean
      source:
        type: object
  ImageImportModel:
    required:
      - public
      - remoteAlias
      - localAlias
      - server
      - description
    properties:
      public:
        type: boolean
      remoteAlias:
        type: string
        minLength: 1
      localAlias:
        type: string
        minLength: 1
      server:
        type: string
        minLength: 1
      description:
        type: string
        minLength: 1
  ContainerConfigModel:
    properties:
      boot.autostart:
        type: boolean
      boot.autostart.delay:
        type: integer
      boot.autostart.priority:
        type: integer
      limits.cpu:
        type: string
      limits.cpu.allowance:
        type: string
      limits.cpu.priority :
        type: integer
      limits.memory:
        type: string
      limits.memory.enforce:
        type: string
      limits.memory.swap:
        type: boolean
      limits.memory.swap.priority:
        type: integer
      security.nesting:
        type: boolean
      security.privileged:
        type: boolean
      volatile.base_image:
        type: string
      volatile.last_state.idmap:
        type: string
      volatile.last_state.power:
        type: string
      limits.memory.enforce:
        type: string
  AuthResponse:
    properties:
      token:
        type: string
      error:
        type: string
  ServerInfoResponse:
    required:
      - serverinfo
    properties:
      serverinfo:
        type: object
  ContainersResponse:
    required:
      - containers
    properties:
      containers:
        type: array
        items:
          $ref: "#/definitions/ContainerModel"
  ProfilesResponse:
    required:
      - profiles
    properties:
      profiles:
        type: array
        items:
          $ref: "#/definitions/ProfileModel"
  ImagesResponse:
    required:
      - images
    properties:
      images:
        type: array
        items:
          $ref: "#/definitions/ImageModel"
  RemoteImagesResponse:
    required:
      - images
    properties:
      images:
        type: array
        items:
          $ref: "#/definitions/RemoteImageModel"
  ServersResponse:
    required:
      - servers
    properties:
      servers:
        type: array
        items:
          $ref: "#/definitions/ServerModel"
  ImageAliasesResponse:
    required:
      - aliases
    properties:
      aliases:
        type: array
        items:
          $ref: "#/definitions/ImageAliasModel"
  ContainerResponse:
    required:
      - container
    properties:
      container:
        $ref: "#/definitions/ContainerModel"
  CreateResponse:
    required:
      - message
    properties:
      message:
        type: string
  DeleteResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string